[FIX] fixed a whole lot of issues with payment requisition, atp, service order and stock move
