+++ topline/data/data.xml
@@ -1,21 +1,6 @@
 <odoo>
 	<data noupdate="0">
 
-		<record id="seq_store_request" model="ir.sequence">
-			<field name="name">Store Request</field>
-			<field name="code">store.request</field>
-			<field name="prefix">SR/%(year)s/</field>
-			<field name="padding">3</field>
-		</record>
-
-		<record id="seq_service_order" model="ir.sequence">
-			<field name="name">Service Order</field>
-			<field name="code">service.order</field>
-			<field name="prefix">SRV/%(year)s/</field>
-			<field name="padding">3</field>
-			<field name="company_id" eval="False"/>
-		</record>
-
 		<record id="seq_project_code" model="ir.sequence">
 			<field name="name">Project Code</field>
 			<field name="code">project.code</field>
@@ -60,14 +45,6 @@
 			<field name="name">Receiving Inspection Report</field>
 			<field name="code">receiving.inspection.report</field>
 			<field name="prefix">RIR/%(year)s/</field>
-			<field name="padding">3</field>
-			<field name="company_id" eval="False"/>
-		</record>
-
-		<record id="seq_atp_form" model="ir.sequence">
-			<field name="name">Authorization to Purchase</field>
-			<field name="code">atp.form</field>
-			<field name="prefix">ATP/%(year)s/</field>
 			<field name="padding">3</field>
 			<field name="company_id" eval="False"/>
 		</record>
@@ -161,51 +138,5 @@
 		</record>
 
 	</data>
-	<data>
-		<record id="stock_location_employee" model="stock.location">
-			<field name="name">Employees</field>
-			<field name="location_id" ref="stock.stock_location_locations_partner"/>
-			<field name="usage">customer</field>
-			<field name="company_id"></field>
-		</record>
-
-		<!-- Removed during migration: put back later -->
-
-		<record id="stock_picking_type_emp" model="stock.picking.type">
-			<field name="name">Staff Store Requests</field>
-			<field name="code">internal</field>
-			<field name="use_create_lots">0</field>
-			<field name="use_existing_lots">1</field>
-			<field name="sequence_id" ref="seq_store_request"/>
-			<field name="sequence_code">INT</field>
-			<field name="default_location_src_id" ref="stock_location_employee"/>
-			<field name="default_location_dest_id" ref="stock_location_employee"/>
-		</record>
-
-		<record model="ir.actions.server" id="pick_stock_view">
-			<field name="name">Stock Picking Server Action</field>
-			<field name="model_id" ref="model_stock_picking"/>
-			<field name="code">
-		        if record.picking_type_id.name == "Staff Store Requests":
-		            action = {
-		                "type": "ir.actions.act_window",
-		                "view_id": env.ref('topline.store_req_action1_window').id,
-		                "view_mode": "form",
-		                "res_model": "stock.picking",
-		                'res_id': record.id,
-		            }
-			</field>
-		</record>
-
-		<record id="product_template_service" model="product.template">
-			<field name="name">Service Request</field>
-			<field name="categ_id" ref="product.cat_expense" />
-			<field name="uom_id" ref="uom.product_uom_unit" />
-			<field name="uom_po_id" ref="uom.product_uom_unit" />
-			<field name="detailed_type">service</field>
-			<field name="invoice_policy">order</field>
-			<field name="purchase_ok">True</field>
-		</record>
-
-	</data>
+	
 </odoo>
