+++ topline/models/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from . import models
-from . import stock
+from . import brand_type
 from . import retirement
 from . import request_log
 from . import account_payment
@@ -10,7 +10,6 @@ from . import maintenance_request
 from . import employee_requisition
 from . import project
 from . import employee_cash_advance
-# from . import account_move
 from . import crm_lead
 from . import project_task
 from . import account_analytic
@@ -20,7 +19,6 @@ from . import hr_leave
 from . import daily_checklist
 from . import exit_form
 from . import salary_advance
-# from . import payment_requisition
-# from . import fleet_vehicle_cost
-# from . import account_asset
-# from . import service_order
+from . import account_move
+from . import hr_expense
+from . import brand_type
