+++ topline/security/ir.model.access.csv
@@ -91,8 +91,6 @@ access_project_monitoring_sheet,project.monitoring.sheet,model_project_monitorin
 access_client_feedback_log,client.feedback.log,model_client_feedback_log,base.group_user,1,1,1,1
 access_client_complaints_log,client.complaints.log,model_client_complaints_log,base.group_user,1,1,1,1
 access_qaqc_vendors_facility_audit_checklist,qaqc.vendors.facility.audit.checklist,model_qaqc_vendors_facility_audit_checklist,base.group_user,1,1,1,1
-access_atp_form,atp.form,model_atp_form,base.group_user,1,1,1,1
-access_atp_form_lines,atp.form.lines,model_atp_form_lines,base.group_user,1,1,1,1
 access_travel_report,travel.report,model_travel_report,base.group_user,1,1,1,1
 access_stationery_report,stationery.report,model_stationery_report,base.group_user,1,1,1,1
 access_salary_advance_form,salary.advance.form,model_salary_advance_form,base.group_user,1,1,1,1
@@ -106,8 +104,6 @@ access_site_time_sheet,site.time.sheet,model_site_time_sheet,base.group_user,1,1
 access_ems_monitoring,ems.monitoring,model_ems_monitoring,base.group_user,1,1,1,1
 access_ems_kpi,ems.kpi,model_ems_kpi,base.group_user,1,1,1,1
 access_fumigation_schedule,fumigation.schedule,model_fumigation_schedule,base.group_user,1,1,1,1
-access_stock_picking_rejected,stock.picking.rejected,model_stock_picking_rejected,base.group_user,1,1,1,1
-access_stock_rejection_reason,stock.rejection.reason,model_stock_rejection_reason,base.group_user,1,1,1,1
 access_brand_type,brand.type,model_brand_type,base.group_user,1,1,1,1
 access_daily_checklist_equipment,daily.checklist.equipment,model_daily_checklist_equipment,base.group_user,1,1,1,1
 access_daily_checklist_power_on_line,daily.checklist.power.off.line,model_daily_checklist_power_on_line,base.group_user,1,1,1,1
