+++ topline_atp/data/ir_sequence.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <data>
+    
+        <record id="seq_atp_form" model="ir.sequence">
+			<field name="name">Authorization to Purchase</field>
+			<field name="code">atp.form</field>
+			<field name="prefix">ATP/%(year)s/</field>
+			<field name="padding">3</field>
+			<field name="company_id" eval="False"/>
+		</record>
+    
+    </data>
+    
+
+</odoo>
