+++ topline_atp/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_atp_form,atp.form,model_atp_form,base.group_user,1,1,1,1
+access_atp_form_lines,atp.form.lines,model_atp_form_lines,base.group_user,1,1,1,1
