+++ topline_payment_requisition/__manifest__.py
@@ -18,7 +18,6 @@
     'depends': [
         'base',
         'topline',
-        'topline_service_order',
     ],
 
     'license': 'LGPL-3',
