+++ topline_service_order/data/ir_sequence.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <data noupdate="0">
+
+        <record id="seq_service_order" model="ir.sequence">
+            <field name="name">Service Order</field>
+            <field name="code">service.order</field>
+            <field name="prefix">SRV/%(year)s/</field>
+            <field name="padding">3</field>
+            <field name="company_id" eval="False"/>
+        </record>
+
+    </data>
+
+
+</odoo>
