+++ topline_service_order/data/service_order_data.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <data noupdate="0">
+    
+        <record id="product_template_service" model="product.template">
+			<field name="name">Service Request</field>
+			<field name="categ_id" ref="product.cat_expense" />
+			<field name="uom_id" ref="uom.product_uom_unit" />
+			<field name="uom_po_id" ref="uom.product_uom_unit" />
+			<field name="detailed_type">service</field>
+			<field name="invoice_policy">order</field>
+			<field name="purchase_ok">True</field>
+			<field name="purchase_method">purchase</field>
+		</record>
+    
+    </data>
+    
+
+</odoo>
