+++ topline_service_order/models/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 
 from . import service_order
+from . import payment_requistion
+from . import purchase_order
