+++ topline_service_order/models/purchase_order.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+from datetime import date
+from odoo import models, fields, api
+
+
+class PurchaseOrder(models.Model):
+    _inherit = 'purchase.order'
+
+    service_order_id = fields.Many2one('service.order', string='Service Order')
