+++ topline_store_request/data/store_request_data.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <data noupdate="0">
+    
+		<record id="stock_location_employee" model="stock.location">
+			<field name="name">Employees</field>
+			<field name="location_id" ref="stock.stock_location_locations_partner"/>
+			<field name="usage">customer</field>
+			<field name="company_id"></field>
+		</record>
+
+		<record id="stock_picking_type_emp" model="stock.picking.type">
+			<field name="name">Staff Store Requests</field>
+			<field name="code">internal</field>
+			<field name="use_create_lots">0</field>
+			<field name="use_existing_lots">1</field>
+			<field name="sequence_id" ref="seq_store_request"/>
+			<field name="sequence_code">INT</field>
+			<field name="default_location_src_id" ref="stock_location_employee"/>
+			<field name="default_location_dest_id" ref="stock_location_employee"/>
+		</record>
+
+		<record model="ir.actions.server" id="pick_stock_view">
+			<field name="name">Stock Picking Server Action</field>
+			<field name="model_id" ref="stock.model_stock_picking"/>
+			<field name="code">
+		        if record.picking_type_id.name == "Staff Store Requests":
+		            action = {
+		                "type": "ir.actions.act_window",
+		                "view_id": env.ref('topline.store_req_action1_window').id,
+		                "view_mode": "form",
+		                "res_model": "stock.picking",
+		                'res_id': record.id,
+		            }
+			</field>
+		</record>
+    
+    </data>
+
+</odoo>
